Cleaned data for running Main.ipynb (5000 samples)
Each tweet has been vectorized; with 0 and 1. There are note all of the data in here, just 5000 samples. Because GitHub has volume limit. To see how this file had been achieved, look at 'New Advanced Processing Method' folder in milestone2. Please note that we just put 5000 samples, because of 25MB volume limit of GitHub. If you want to see all 100,000 samples, email chix4030@gmail.com.
Please note that we are not encoding the name of the classes, because we want you to see them; we encode them in Main.ipyny easily.